fixed some function signature issues
