+++ setup.py
@@ -10,7 +10,7 @@ long_description = (this_directory / "README.md").read_text()
 
 setup(
     name='paramiko-mock',
-    version='0.0.3',
+    version='0.0.4',
     description='Mock for paramiko library',
     author='Caio Cominato, Daniel Beaudry',
     author_email='caiopetrellicominato@gmail.com',
